Encoder now gives data, need to check conversion

+++ src/org/usfirst/frc/team1038/robot/Encoder1038.java
@@ -38,7 +38,7 @@ public class Encoder1038 {
 		return encoder.getDistance();
 	}
 	
-	private double findDistancePerPulse(int countsPerRevolution, double wheelDiameter)
+	private double findDistancePerPulse(double countsPerRevolution, double wheelDiameter)
 	{		
 		return (1 / countsPerRevolution ) * (wheelDiameter * Math.PI);
 	}
